requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
pdfplumber>=0.11.0
google-genai>=1.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.2.0
aiofiles>=23.2.1
